Moved size.go to size_os.go files except for size_windows.go because windows doesn't have the specified syscalls, and the package wouldn't compile on windows as a result.
Added runtime.GOOS != "windows" check for color printing;
win32 consoles (cmd prompt is one of these) don't have
support for terminal color codes.
