+++ term/size_netbsd.go
@@ -1,0 +1,1 @@
+size_linux.go
