+++ term/size_openbsd.go
@@ -1,0 +1,1 @@
+size_linux.go
